prioritized UI Updates instead of slow API, more structured code, moved working pages to /app
