--- src/app/(pages)/test/page.tsx
@@ -1,3 +1,0 @@
-export default function TestPage() {
-  return <div>{/* <Project /> */}</div>;
-}
