+++ src/app/api/(kanban)/task/update/route.ts
@@ -1,4 +1,3 @@
-import { getUserData } from "@/actions/getUserData";
 import taskConfilctResolver from "@/actions/taskConfilctResolver";
 import prisma from "@/db";
 import { authOptions } from "@/lib/auth";
@@ -8,46 +7,36 @@ import { NextRequest, NextResponse } from "next/server";
 
 export async function PUT(request: NextRequest) {
   const { id, updateValue } = await request.json();
-
   if (!id || !updateValue) {
     return NextResponse.json({ message: "Invalid Request" });
   }
-
   const updates = await taskSchema.parseAsync(updateValue);
 
   if (!updates) {
     return NextResponse.json({ message: "Invalid Request" });
   }
 
-  const session = await getServerSession(authOptions);
-
   try {
-    const user = await getUserData(session);
-    if (!user) {
-      return NextResponse.json({ message: "User not found" }, { status: 404 });
+    const session = await getServerSession(authOptions);
+    if (!session?.user) {
+      return NextResponse.json({ message: "Please sign in first to continue" }, { status: 401 });
     }
 
-    const task = await prisma.task.findUnique({
-      where: {
-        id,
-      },
-    });
-    if (!task) {
-      return NextResponse.json({ message: "Task doesn't exist" }, { status: 404 });
-    }
-
-    const newTask = await prisma.task.update({
+    const task = await prisma.task.update({
       where: {
         id,
       },
       data: updates,
     });
-
-    if (updates.position) {
-      taskConfilctResolver(newTask.categoryId);
+    if (!task) {
+      return NextResponse.json({ message: "Task doesn't exist" }, { status: 404 });
     }
 
-    return NextResponse.json({ message: "Task updated", task: newTask });
+    if (updates.position) {
+      taskConfilctResolver(task.categoryId);
+    }
+
+    return NextResponse.json({ message: "Task updated", task: task });
   } catch (error) {
     console.log(error);
     return NextResponse.json({ message: "Something went wrong" }, { status: 500 });
