+++ src/app/app/(kanban)/inbox/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Board from "@/components/dnd-new/Board";
+import Board from "@/components/dnd/Board";
 
 export default function Inbox() {
   return <Board heading="Inbox" />;
