+++ src/app/app/(kanban)/today/page.tsx
@@ -1,0 +1,5 @@
+import Board from "@/components/dnd/Board";
+
+export default async function Today() {
+  return <Board heading="Today" />;
+}
