--- src/app/app/today/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import Board from "@/components/dnd-new/Board";
-
-export default function Today() {
-  return <Board heading="Today" />;
-}
