+++ src/middleware.ts
@@ -72,11 +72,9 @@ export const config = {
     "/app/inbox",
     "/app/today",
     "/labels",
-    "/onboard/profile",
-    "/onboard/use-case",
-    "/projects",
-    "/projects/:projectId",
-    "/test",
+    "/app/onboard/profile",
+    "/app/onboard/use-case",
+    "/app/projects",
     "/sign-in",
     "/sign-up",
     "/forgot-password",
